Move camera movement into camtest
