+++ src/bin/test_camera/main.rs
@@ -1,8 +1,35 @@
 use bevy::prelude::*;
-use bounded_planet::camera::{CameraBP, CameraBPPlugin, UniversalGeometry};
+use bounded_planet::camera::*;
+
+/// The threshold for horizontal cursor-activated [`CameraBP`] movement.
+///
+/// Is a proportion of the window size. So, if this is `0.05`, then the cursor
+/// must be within 5% of the window size to either the left or right edge to
+/// trigger this threshold.
+const CURSOR_EDGE_H_THRESHOLD: f32 = 0.05;
+/// The threshold for vertical cursor-activated [`CameraBP`] movement.
+///
+/// Is a proportion of the window size. So, if this is `0.05`, then the cursor
+/// must be within 5% of the window size to either the top or bottom edge to
+/// trigger this threshold.
+const CURSOR_EDGE_V_THRESHOLD: f32 = 0.05;
+
+/// The stage at which the [`CameraBP`] cache is either updated or used to fill
+/// in the action cache now.
+const CAM_CACHE_UPDATE: &'static str = "push_cam_update";
+
+#[derive(Copy, Clone)]
+struct IsActionCacheDirty(bool);
+
+impl Default for IsActionCacheDirty {
+    fn default() -> Self {
+        IsActionCacheDirty(true)
+    }
+}
 
 fn main() {
     App::build()
+        .init_resource::<IsActionCacheDirty>()
         .add_resource(Msaa { samples: 4 })
         .add_default_plugins()
         .add_plugin(CameraBPPlugin {
@@ -13,6 +40,9 @@ fn main() {
             ..Default::default()
         })
         .add_startup_system(setup.system())
+        .add_system_to_stage(stage::EVENT_UPDATE, act_camera_on_window_edge.system())
+        .add_stage_after(stage::EVENT_UPDATE, CAM_CACHE_UPDATE)
+        .add_system_to_stage(CAM_CACHE_UPDATE, use_or_update_action_cache.system())
         .run();
 }
 
@@ -58,4 +88,52 @@ fn setup(
             rotation: Rotation::from_rotation_xyz(-0.75, 2.7, 0.0),
             ..Default::default()
         }).with(CameraBP);
+}
+
+/// Pushes camera actions based upon mouse movements near the window edge.
+fn act_camera_on_window_edge(
+    wins: Res<Windows>,
+    mut dirty: ResMut<IsActionCacheDirty>,
+    pos: Res<Events<CursorMoved>>,
+    mut cams: ResMut<Events<CameraBPAction>>
+) {
+    dirty.0 = false;
+
+    if let Some(e) = pos.get_reader().find_latest(&pos, | e | e.id.is_primary()) {
+        let (mouse_x, mouse_y) = (e.position.x(), e.position.y());
+        let window = wins.get(e.id).expect("Couldn't get primary window.");
+        let (window_x, window_y) = (window.width as f32, window.height as f32);
+        dirty.0 = true;
+
+        if mouse_x / window_x <= CURSOR_EDGE_H_THRESHOLD {
+            cams.send(CameraBPAction::MoveLeft(None));
+        }
+
+        if 1.0 - mouse_x / window_x <= CURSOR_EDGE_H_THRESHOLD {
+            cams.send(CameraBPAction::MoveRight(None));
+        }
+
+        if mouse_y / window_y <= CURSOR_EDGE_V_THRESHOLD {
+            cams.send(CameraBPAction::MoveBack(None));
+        }
+
+        if 1.0 - mouse_y / window_y <= CURSOR_EDGE_V_THRESHOLD {
+            cams.send(CameraBPAction::MoveForward(None));
+        }
+    }
+}
+
+/// Depending on `dirty`, either update the local `cache` or fill the event
+/// queue for [`CameraBPAction`] with the locally cached copy.
+fn use_or_update_action_cache(
+    mut cache: Local<Vec<CameraBPAction>>,
+    mut cams: ResMut<Events<CameraBPAction>>,
+    dirty: Res<IsActionCacheDirty>
+) {
+    if dirty.0 {
+        *cache = cams.get_reader().iter(&cams).copied().collect();
+        cache.dedup_by(| l, r | l.both_same_signal(r));
+    } else {
+        cams.extend(cache.iter().copied())
+    }
 }
